fix: icon optimization, input type fixes , added story book
